+++ pagerduty_integration_test.go
@@ -12,7 +12,7 @@ func TestCreatePagerDutyIntegration(t *testing.T) {
 	teardown := setup()
 	defer teardown()
 
-	mux.HandleFunc("/v2/integration", verifyRequest(t, "POST", http.StatusOK, nil, "integration/create_pd_success.json"))
+	mux.HandleFunc("/v2/integration", verifyRequest(t, "POST", true, http.StatusOK, nil, "integration/create_pd_success.json"))
 
 	result, err := client.CreatePagerDutyIntegration(&integration.PagerDutyIntegration{
 		Type: "PagerDuty",
@@ -25,7 +25,7 @@ func TestGetPagerDutyIntegration(t *testing.T) {
 	teardown := setup()
 	defer teardown()
 
-	mux.HandleFunc("/v2/integration/id", verifyRequest(t, "GET", http.StatusOK, nil, "integration/create_pd_success.json"))
+	mux.HandleFunc("/v2/integration/id", verifyRequest(t, "GET", true, http.StatusOK, nil, "integration/create_pd_success.json"))
 
 	result, err := client.GetPagerDutyIntegration("id")
 	assert.NoError(t, err, "Unexpected error getting integration")
@@ -36,7 +36,7 @@ func TestUpdatePagerDutyIntegration(t *testing.T) {
 	teardown := setup()
 	defer teardown()
 
-	mux.HandleFunc("/v2/integration/id", verifyRequest(t, "PUT", http.StatusOK, nil, "integration/create_pd_success.json"))
+	mux.HandleFunc("/v2/integration/id", verifyRequest(t, "PUT", true, http.StatusOK, nil, "integration/create_pd_success.json"))
 
 	result, err := client.UpdatePagerDutyIntegration("id", &integration.PagerDutyIntegration{
 		Type: "PagerDuty",
@@ -49,7 +49,7 @@ func TestDeletePagerDutyIntegration(t *testing.T) {
 	teardown := setup()
 	defer teardown()
 
-	mux.HandleFunc("/v2/integration/id", verifyRequest(t, "DELETE", http.StatusNoContent, nil, ""))
+	mux.HandleFunc("/v2/integration/id", verifyRequest(t, "DELETE", true, http.StatusNoContent, nil, ""))
 
 	err := client.DeletePagerDutyIntegration("id")
 	assert.NoError(t, err, "Unexpected error creating integration")